pyproject.toml
[project]
name = "pi-webcam"
version = "0.1.0"
description = "Wildlife camera for Raspberry Pi Zero 2 W with live streaming and timelapse capture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.1.0",
    "pillow>=10.0.0",
]

[project.scripts]
pi-webcam = "pi_webcam.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pi_webcam"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = [
    "B008",   # Depends() in function defaults — standard FastAPI pattern
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["PIL", "PIL.*"]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "mypy>=1.10.0",
    "ruff>=0.5.0",
]
